chore: moves shared features to desktop
