+++ packages/desktop/electron/lib/menu.js
@@ -1,6 +1,6 @@
 import { app, ipcMain, Menu, shell } from 'electron'
-import { DISCORD_URL, DOCUMENTATION_URL, FAQ_URL, ISSUE_REPORT_URL } from 'shared/lib/contexts/settings/constants'
-import features from 'shared/features/features'
+import { DISCORD_URL, DOCUMENTATION_URL, FAQ_URL, ISSUE_REPORT_URL } from '@contexts/settings/constants'
+import features from '@features/features'
 import { closeAboutWindow, getOrInitWindow, openAboutWindow } from '../main'
 import { menuState } from './menuState'
