+++ packages/desktop/tsconfig.json
@@ -11,7 +11,7 @@
             "@contexts/*": ["../shared/lib/contexts/*"],
             "@components": ["./components"],
             "@core/*": ["../shared/lib/core/*"],
-            "@features/*": ["../shared/features/*"],
+            "@features/*": ["./features/*"],
             "@lib/*": ["../shared/lib/*"],
             "@ui": ["../shared/components"],
             "@views": ["./views"]
