+++ packages/desktop/views/dashboard/settings/views/security/Security.svelte
@@ -3,7 +3,7 @@
     import { SecuritySettingsRoute } from '@core/router'
     import { HR } from 'shared/components'
     import { AppLock, ChangePassword, ChangePincode, DeleteProfile, ExportStronghold } from './'
-    import features from 'shared/features/features'
+    import features from '@features/features'
 
     const settings: {
         component: unknown
